# packages/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CustomerType": {
        "enum": [
          "MEMBER",
          "GUEST"
        ],
        "title": "CustomerType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemberReservationCreate": {
        "properties": {
          "court_id": {
            "format": "uuid",
            "title": "Court Id",
            "type": "string"
          },
          "duration_minutes": {
            "$ref": "#/components/schemas/AppDuration"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "court_id",
          "starts_at",
          "duration_minutes"
        ],
        "title": "MemberReservationCreate",
        "type": "object"
      },
...
      "MyReservationOut": {
        "properties": {
          "can_cancel": {
            "title": "Can Cancel",
            "type": "boolean"
          },
          "cancel_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Deadline"
          },
          "cancel_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CancelReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "club": {
            "$ref": "#/components/schemas/ClubBrief"
          },
          "court": {
            "$ref": "#/components/schemas/CourtBrief"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus"
          },
          "total_price": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
            "title": "Total Price",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "starts_at",
          "ends_at",
          "duration_minutes",
          "total_price",
          "cancel_reason",
          "created_at",
          "club",
          "court",
          "can_cancel",
          "cancel_deadline"
        ],
        "title": "MyReservationOut",
        "type": "object"
      },
...
      "Page_ReservationOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReservationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ReservationOut]",
        "type": "object"
      },
...
      "ReservationCreate": {
        "description": "Reserva cargada por el staff: de un socio (`membership_id`) o de un invitado.",
        "properties": {
          "court_id": {
            "format": "uuid",
            "title": "Court Id",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "guest_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          },
          "guest_phone": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Phone"
          },
          "membership_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Membership Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price_override": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Override"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "court_id",
          "starts_at",
          "ends_at"
        ],
        "title": "ReservationCreate",
        "type": "object"
      },
      "ReservationGridOut": {
        "properties": {
          "close_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Time"
          },
          "courts": {
            "items": {
              "$ref": "#/components/schemas/GridCourtOut"
            },
            "title": "Courts",
            "type": "array"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "open_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Time"
          },
          "slot_minutes": {
            "title": "Slot Minutes",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "date",
          "timezone",
          "open_time",
          "close_time",
          "slot_minutes",
          "courts"
        ],
        "title": "ReservationGridOut",
        "type": "object"
      },
      "ReservationOut": {
        "properties": {
          "cancel_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CancelReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "court_id": {
            "format": "uuid",
            "title": "Court Id",
            "type": "string"
          },
          "court_name": {
            "title": "Court Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_name": {
            "title": "Customer Name",
            "type": "string"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "customer_type": {
            "$ref": "#/components/schemas/CustomerType"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "member_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Number"
          },
          "membership_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Membership Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "paid_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid Amount",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ReservationSource"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus"
          },
          "total_price": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Price",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "court_id",
          "court_name",
          "customer_type",
          "customer_name",
          "customer_phone",
          "user_id",
          "membership_id",
          "member_number",
          "status",
          "source",
          "starts_at",
          "ends_at",
          "duration_minutes",
          "total_price",
          "paid_amount",
          "notes",
          "confirmed_at",
          "cancelled_at",
          "cancel_reason",
          "created_at"
        ],
        "title": "ReservationOut",
        "type": "object"
      },
...
      "ReservationStatus": {
        "enum": [
          "pending",
          "confirmed",
          "cancelled",
          "completed"
        ],
        "title": "ReservationStatus",
        "type": "string"
      },
      "ReservationUpdate": {
        "description": "Solo se modifican los campos enviados. Reprogramar exige `starts_at` y `ends_at` juntos.\nSi cambia la cancha o la duración, el precio se recalcula salvo `price_override`.",
        "properties": {
          "court_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price_override": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Override"
          },
          "starts_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          }
        },
        "title": "ReservationUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ClubSystem API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/reservations": {
      "get": {
        "description": "Reservas de un día (`date`) o de un rango (`from`/`to`), en días locales del club.",
        "operationId": "list_reservations_api_v1_admin_reservations_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReservationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "court_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Court Id"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "starts_at",
              "enum": [
                "starts_at",
                "-starts_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReservationOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reservations",
        "tags": [
          "Admin: reservas"
        ]
      },
      "post": {
        "operationId": "create_reservation_api_v1_admin_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Reservation",
        "tags": [
          "Admin: reservas"
        ]
      }
    },
    "/api/v1/admin/reservations/export": {
      "get": {
        "operationId": "export_reservations_api_v1_admin_reservations_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Reservations",
        "tags": [
          "Admin: reservas"
        ]
      }
    },
    "/api/v1/admin/reservations/grid": {
      "get": {
        "description": "Canchas activas con las reservas no canceladas del día, para la grilla.",
        "operationId": "reservation_grid_api_v1_admin_reservations_grid_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationGridOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reservation Grid",
        "tags": [
          "Admin: reservas"
        ]
      }
    },
    "/api/v1/admin/reservations/{reservation_id}": {
      "get": {
        "operationId": "get_reservation_api_v1_admin_reservations__reservation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reservation",
        "tags": [
          "Admin: reservas"
        ]
      },
      "patch": {
        "operationId": "update_reservation_api_v1_admin_reservations__reservation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Reservation",
        "tags": [
          "Admin: reservas"
        ]
      }
    },
    "/api/v1/admin/reservations/{reservation_id}/cancel": {
      "post": {
        "operationId": "cancel_reservation_api_v1_admin_reservations__reservation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Reservation",
        "tags": [
          "Admin: reservas"
        ]
      }
    },
    "/api/v1/admin/reservations/{reservation_id}/confirm": {
      "post": {
        "operationId": "confirm_reservation_api_v1_admin_reservations__reservation_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Reservation",
        "tags": [
          "Admin: reservas"
        ]
      }
    },
...
    "/api/v1/mobile/clubs/{club_id}/reservations": {
      "post": {
        "description": "Queda pendiente hasta que la confirme el staff.",
        "operationId": "create_reservation_api_v1_mobile_clubs__club_id__reservations_post",
        "parameters": [
          {
            "in": "path",
            "name": "club_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Club Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberReservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Reservation",
        "tags": [
          "App: reservas"
        ]
      }
    },
    "/api/v1/mobile/clubs/{club_id}/reservations/{reservation_id}/cancel": {
      "post": {
        "description": "Pendientes siempre; confirmadas hasta `member_cancel_notice_hours` antes del inicio.",
        "operationId": "cancel_reservation_api_v1_mobile_clubs__club_id__reservations__reservation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "club_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Club Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Reservation",
        "tags": [
          "App: reservas"
        ]
      }
    },
...
    "/api/v1/mobile/reservations": {
      "get": {
        "description": "Reservas propias, de todos los clubes del usuario.",
        "operationId": "list_my_reservations_api_v1_mobile_reservations_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "upcoming",
              "enum": [
                "upcoming",
                "past"
              ],
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MyReservationOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Reservations",
        "tags": [
          "App: reservas"
        ]
      }
    },
    "/api/v1/mobile/reservations/{reservation_id}": {
      "get": {
        "operationId": "get_my_reservation_api_v1_mobile_reservations__reservation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Reservation",
        "tags": [
          "App: reservas"
        ]
      }
    },
...
  }
}